+++ lib/handler/handle-json.js
@@ -1,0 +1,10 @@
+import { createHandlerFactory } from './factory'
+import { getUniversalCtx } from './get-context'
+
+const identity = (data) => data
+
+export const createJsonHandler = createHandlerFactory(
+  identity,
+  identity,
+  getUniversalCtx
+)
