+++ lib/handler/index.js
@@ -1,1 +1,1 @@
-export * from './factory'
+export * from './handle-json'
